requirements.txt

fastapi==0.121.0
uvicorn==0.38.0
jinja2==3.1.6
python-dotenv==1.2.1
python-multipart==0.0.20
ollama==0.6.0
